Optimize data types for market event structures
- Add explicit byte backing (: byte) to all 13 domain enums to reduce
  serialized size and memory per event (MarketEventType, AggressorSide,
  OrderBookSide, DepthOperation, IntegritySeverity, MarketState,
  MarketEventTier, DepthIntegrityKind, OptionRight, OptionStyle,
  InstrumentType, AggregateTimeframe x2)
- Narrow OrderBookLevel.Level and depth Position fields from int to
  ushort (max 65,535 levels far exceeds any real order book depth)
- Narrow MarketEvent.SchemaVersion and CanonicalizationVersion from
  int to byte (max 255 versions is sufficient)
- Remove tautological negative-value checks on unsigned fields
- Remove obsolete negative-level test case (ushort can't be negative)
- Update all callers across source, tests, and benchmarks

+++ benchmarks/MarketDataCollector.Benchmarks/CollectorBenchmarks.cs
@@ -125,7 +125,7 @@ public class DepthCollectorBenchmarks
             _insertUpdates[i] = new MarketDepthUpdate(
                 Timestamp: DateTimeOffset.UtcNow.AddMilliseconds(i),
                 Symbol: "SPY",
-                Position: pos > 0 ? random.Next(0, Math.Min(pos, 49)) : 0,
+                Position: (ushort)(pos > 0 ? random.Next(0, Math.Min(pos, 49)) : 0),
                 Operation: DepthOperation.Insert,
                 Side: side,
                 Price: side == OrderBookSide.Bid ? 450m - pos * 0.01m : 450.01m + pos * 0.01m,
@@ -160,7 +160,7 @@ public class DepthCollectorBenchmarks
             _mixedUpdates[i] = new MarketDepthUpdate(
                 Timestamp: DateTimeOffset.UtcNow.AddMilliseconds(i),
                 Symbol: "SPY",
-                Position: pos,
+                Position: (ushort)pos,
                 Operation: op,
                 Side: side,
                 Price: side == OrderBookSide.Bid ? 450m - pos * 0.01m : 450.01m + pos * 0.01m,
